move FetcherError outside
