+++ src/helpers/axios.ts
@@ -1,60 +1,6 @@
 import { FetcherFunction } from '../types';
 import type Axios from 'axios';
-import type { AxiosError } from 'axios';
-
-type FetcherErrorConstructorInput = Pick<
-  AxiosError,
-  'code' | 'request' | 'response' | 'isAxiosError' | 'message'
->;
-
-interface FetcherErrorInterface extends FetcherErrorConstructorInput {
-  status?: number;
-  code?: string;
-}
-
-export const isFetcherError = (error: unknown): error is FetcherError => {
-  return error instanceof FetcherError;
-};
-
-export class FetcherError extends Error implements FetcherErrorInterface {
-  code;
-  status;
-  request;
-  response;
-  isAxiosError = false;
-
-  public constructor({
-    code,
-    request,
-    response,
-    isAxiosError,
-    message,
-  }: FetcherErrorConstructorInput) {
-    super();
-
-    // Maintains proper stack trace for where our error was thrown (only available on V8)
-    if (Error.captureStackTrace) {
-      Error.captureStackTrace(this, FetcherError);
-    }
-
-    this.name = this.constructor.name;
-    this.code = code;
-    this.request = request;
-    this.isAxiosError = isAxiosError;
-
-    // no response on AxiosError
-    if (!response) {
-      this.message = message;
-      return;
-    }
-
-    this.response = response;
-
-    const { data, status } = response;
-    this.status = status;
-    this.message = data?.error || message;
-  }
-}
+import { FetcherError } from './misc';
 
 export const constructFetcher =
   (axios: typeof Axios): FetcherFunction =>
@@ -63,7 +9,7 @@ export const constructFetcher =
       const { data } = await axios.request(params);
 
       return data;
-    } catch (error: AxiosError | any) {
+    } catch (error: any) {
       if (axios.isAxiosError(error)) {
         // attach a more meaningful message and other props
         throw new FetcherError(error);
