+++ src/helpers/misc.ts
@@ -6,6 +6,8 @@ import {
 } from 'ethers';
 import { assert } from 'ts-essentials';
 
+import type { AxiosError } from 'axios';
+
 export type ContractWithMethod<T extends string> = Contract & {
   readonly [method in T]: ContractFunction;
 } & {
@@ -56,3 +58,57 @@ export const constructSearchString = <U extends Record<string, unknown>>(
   // returns empty string or `?${string}`
   return queryString && `?${queryString}`;
 };
+
+type FetcherErrorConstructorInput = Pick<
+  AxiosError,
+  'code' | 'request' | 'response' | 'isAxiosError' | 'message'
+>;
+
+interface FetcherErrorInterface extends FetcherErrorConstructorInput {
+  status?: number;
+  code?: string;
+}
+
+export const isFetcherError = (error: unknown): error is FetcherError => {
+  return error instanceof FetcherError;
+};
+
+export class FetcherError extends Error implements FetcherErrorInterface {
+  code;
+  status;
+  request;
+  response;
+  isAxiosError = false;
+
+  public constructor({
+    code,
+    request,
+    response,
+    isAxiosError,
+    message,
+  }: FetcherErrorConstructorInput) {
+    super();
+
+    // Maintains proper stack trace for where our error was thrown (only available on V8)
+    if (Error.captureStackTrace) {
+      Error.captureStackTrace(this, FetcherError);
+    }
+
+    this.name = this.constructor.name;
+    this.code = code;
+    this.request = request;
+    this.isAxiosError = isAxiosError;
+
+    // no response on AxiosError
+    if (!response) {
+      this.message = message;
+      return;
+    }
+
+    this.response = response;
+
+    const { data, status } = response;
+    this.status = status;
+    this.message = data?.error || message;
+  }
+}
